newsletter_project/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
newspaper3k>=0.2.8
sumy>=0.11.0
lxml[html_clean]>=5.0.0
pandas>=2.0.0
nltk>=3.8.0
python-docx>=1.0.0
